moved reminder logic to shift class
